fix: cast recordAiUsage's ambiguous bigint/numeric parameters, add regression coverage

+++ packages/domain/src/index.ts
@@ -115,12 +115,12 @@ export async function recordAiUsage(input: { runId: string } & AiUsage, client?:
        ORDER BY p.effective_from DESC LIMIT 1
      )
      UPDATE agent_runs ar SET
-       ai_usage_status='captured', input_tokens=$2, output_tokens=$3, reasoning_tokens=$4,
-       cache_read_tokens=$5, cache_write_tokens=$6, total_tokens=$7, raw_usage_json=$8,
+       ai_usage_status='captured', input_tokens=$2::bigint, output_tokens=$3::bigint, reasoning_tokens=$4::bigint,
+       cache_read_tokens=$5::bigint, cache_write_tokens=$6::bigint, total_tokens=$7::bigint, raw_usage_json=$8,
        ai_model_price_id=(SELECT id FROM price),
        estimated_cost_usd=(SELECT
-         ($2 * input_usd_per_million + $3 * output_usd_per_million +
-          $5 * cache_read_usd_per_million + $6 * cache_write_usd_per_million) / 1000000
+         ($2::numeric * input_usd_per_million + $3::numeric * output_usd_per_million +
+          $5::numeric * cache_read_usd_per_million + $6::numeric * cache_write_usd_per_million) / 1000000
          FROM price)
      WHERE ar.id=$1 AND ar.ai_usage_status='pending'
      RETURNING ar.*`,
